requirements.txt
# Base requirements for FLUX.1-dev
# For FLUX.1-Kontext-dev, use requirements.kontext.txt
#
# Note: requirements.fluxdev.txt and requirements.kontext.txt are used during Docker build
# This file is kept for compatibility and defaults to FLUX.1-dev settings

# HF ecosystem
safetensors
huggingface-hub
transformers==4.52.4
git+https://github.com/huggingface/diffusers@363d1ab7e24c5ed6c190abb00df66d9edb74383b
sentencepiece
open-clip-torch
accelerate
hf_transfer

# LoRA / extras
lycoris-lora==1.8.3
peft
torchao==0.9.0
optimum-quanto==0.2.4
k-diffusion
einops
kornia
invisible-watermark
prodigyopt
lpips
pytorch-fid
pytorch-wavelets==1.3.0
controlnet-aux==0.0.10
git+https://github.com/jaretburkett/easy_dwpose.git

# Data utils
flatten_json
pyyaml
oyaml
python-dotenv
toml
albumentations==1.4.15
albucore==0.0.16
pydantic
omegaconf
tensorboard
bitsandbytes
python-slugify

# Web/API
fastapi==0.104.1
uvicorn[standard]==0.24.0
gradio
requests==2.31.0
aiohttp
py-consul

# Vision & plotting
opencv-python
matplotlib==3.10.1

# TensorRT and ONNX (these get filtered out by Dockerfile but kept for completeness)
polygraphy
onnx-graphsurgeon
onnxruntime-gpu
tensorrt==10.5.0
tensorrt-cu12-bindings==10.5.0
tensorrt-cu12-libs==10.5.0
cuda-python
nvidia-modelopt

# Additional dependencies
timm

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
